Fix & improvements

+++ ImmuneResponsePredictor/queries/study/study_cohorts_info.sql
@@ -12,5 +12,5 @@ WHERE
  HAI.SUBJECT_ACCESSION = GEM.Biosample.subject_accession AND
  GEM.Biosample.arm_name = GEA.arm_name AND
  GEM.Biosample.study_time_collected > 0 AND
- CAST(REPLACE( GEA.coefficient, 'study_time_collected', '') AS INTEGER) = GEM.Biosample.study_time_collected
+ CAST(REPLACE( GEA.coefficient, 'study_time_collected', '') AS DECIMAL) = GEM.Biosample.study_time_collected
